fix typo

+++ src/EventManager.Core/EventOrganizer/Entities/EventEntity.cs
@@ -19,7 +19,7 @@ namespace EventManager.Core.EventOrganizer.Entities
     [Required(ErrorMessage = "Location of the event is must")]
     public string Location { get; set; }
 
-    public uint? Capacity { get; set; }
+    public ushort? Capacity { get; set; }
 
     public string Country { get; set; }
